Enum mml::core::item::state::abstraction::Abstract

pub enum Abstract<'a> {
    Trait(Trait<'a>),
    Struct(Struct<'a>),
    Enum(Enum<'a>),
    None,
}

The structure Abstract is a enumerate for abstract element types or none.

Variants
Trait(Trait<'a>)Struct(Struct<'a>)Enum(Enum<'a>)None
Methods

impl<'a> Abstract<'a>

fn as_name(&self) -> Option<&InternedString>

The method as_name returns the name of the abstract element or else declare a panic.